OBG-647. Created shared BPMN package for common Flowable utilities
