+++ opba-embedded-starter-tests/src/test/java/de/adorsys/opba/starter/Xs2aSensitiveDataTest.java
@@ -1,6 +1,6 @@
 package de.adorsys.opba.starter;
 
-import de.adorsys.opba.protocol.xs2a.config.flowable.Xs2aObjectMapper;
+import de.adorsys.opba.protocol.bpmnshared.config.flowable.Xs2aObjectMapper;
 import de.adorsys.opba.protocol.xs2a.service.xs2a.context.Xs2aContext;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.Test;
