+++ opba-protocols/protocol-bpmn-shared/README.md
@@ -1,0 +1,3 @@
+# What is this
+
+This is shared definition for **Flowable BPMN** usages by the protocols.
