+++ opba-protocols/protocol-bpmn-shared/src/main/java/de/adorsys/opba/protocol/bpmnshared/config/flowable/JsonCustomSerializer.java
@@ -1,28 +1,28 @@
-package de.adorsys.opba.protocol.xs2a.config.flowable;
+package de.adorsys.opba.protocol.bpmnshared.config.flowable;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import de.adorsys.opba.protocol.api.services.scoped.RequestScopedServicesProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import org.flowable.variable.api.types.ValueFields;
-import org.flowable.variable.service.impl.types.SerializableType;
+import org.flowable.variable.api.types.VariableType;
 
 import java.util.List;
 
 /**
- * JSON serializer for large classes (large resulting strings). Preserves the class name used, so deserialzation
- *  returns the class that was used to serialize data.
+ * JSON serializer for small classes (small resulting strings). Preserves the class name used, so deserialzation
+ * returns the class that was used to serialize data.
  * Data is encrypted using {@link RequestScopedServicesProvider}.
  */
 @RequiredArgsConstructor
-class LargeJsonCustomSerializer extends SerializableType {
+class JsonCustomSerializer implements VariableType {
 
-    static final String JSON = "as_large_json";
+    static final String JSON = "as_json";
 
     private final RequestScopedServicesProvider scopedServicesProvider;
     private final ObjectMapper mapper;
     private final List<String> allowOnlyClassesWithPrefix;
-    private final int minLength;
+    private final int maxLength;
 
     @Override
     public String getTypeName() {
@@ -46,26 +46,29 @@ class LargeJsonCustomSerializer extends SerializableType {
         }
 
         String value = mapper.writeValueAsString(o);
-        return value.length() > minLength;
+        return value.length() < maxLength;
     }
 
     @Override
-    public byte[] serialize(Object o, ValueFields valueFields) {
+    @SneakyThrows
+    public void setValue(Object o, ValueFields valueFields) {
         if (o == null) {
-            return null;
+            valueFields.setBytes(null);
+            return;
         }
 
-        return SerializerUtil.serialize(o, mapper);
+        valueFields.setBytes(SerializerUtil.serialize(o, mapper));
     }
 
     @Override
-    public Object deserialize(byte[] bytes, ValueFields valueFields) {
-        if (null == bytes) {
+    @SneakyThrows
+    public Object getValue(ValueFields valueFields) {
+        if (null == valueFields.getBytes()) {
             return null;
         }
 
         return SerializerUtil.deserialize(
-                bytes,
+                valueFields.getBytes(),
                 mapper,
                 allowOnlyClassesWithPrefix,
                 scopedServicesProvider
