+++ opba-protocols/protocol-bpmn-shared/src/main/java/de/adorsys/opba/protocol/bpmnshared/config/flowable/Xs2aObjectMapper.java
@@ -1,4 +1,4 @@
-package de.adorsys.opba.protocol.xs2a.config.flowable;
+package de.adorsys.opba.protocol.bpmnshared.config.flowable;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.Getter;
