+++ opba-protocols/protocol-bpmn-shared/src/main/java/de/adorsys/opba/protocol/bpmnshared/dto/messages/ProcessResponse.java
@@ -1,4 +1,4 @@
-package de.adorsys.opba.protocol.xs2a.domain.dto.messages;
+package de.adorsys.opba.protocol.bpmnshared.dto.messages;
 
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
