+++ opba-protocols/xs2a-protocol/pom.xml
@@ -24,6 +24,12 @@
             <artifactId>opba-protocol-api</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <!-- Flowable BPMN -->
+        <dependency>
+            <groupId>de.adorsys.opba</groupId>
+            <artifactId>protocol-bpmn-shared</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 
         <!-- javadoc -->
         <dependency>
@@ -40,7 +46,6 @@
         <dependency>
             <groupId>org.jetbrains</groupId>
             <artifactId>annotations</artifactId>
-            <version>${jetbrains.annotations.version}</version>
         </dependency>
 
         <!-- Spring -->
@@ -62,13 +67,6 @@
         <dependency>
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
-        </dependency>
-
-        <!-- Flowable BPMN -->
-        <dependency>
-            <groupId>org.flowable</groupId>
-            <artifactId>flowable-spring-boot-starter-process</artifactId>
-            <version>${flowable.version}</version>
         </dependency>
 
         <!-- XS2A Adapter BEGIN -->
