+++ opba-protocols/xs2a-protocol/src/main/java/de/adorsys/opba/protocol/xs2a/service/validation/Xs2aReportValidationError.java
@@ -1,7 +1,7 @@
 package de.adorsys.opba.protocol.xs2a.service.validation;
 
+import de.adorsys.opba.protocol.bpmnshared.dto.messages.ValidationProblem;
 import de.adorsys.opba.protocol.xs2a.config.protocol.ProtocolConfiguration;
-import de.adorsys.opba.protocol.xs2a.domain.dto.messages.ValidationProblem;
 import de.adorsys.opba.protocol.xs2a.service.ContextUtil;
 import de.adorsys.opba.protocol.xs2a.service.xs2a.context.BaseContext;
 import de.adorsys.opba.protocol.xs2a.service.xs2a.context.LastRedirectionTarget;
