+++ opba-protocols/xs2a-protocol/src/main/java/de/adorsys/opba/protocol/xs2a/service/xs2a/consent/ReportConsentAuthorizationDenied.java
@@ -1,6 +1,6 @@
 package de.adorsys.opba.protocol.xs2a.service.xs2a.consent;
 
-import de.adorsys.opba.protocol.xs2a.domain.dto.messages.ConsentAcquired;
+import de.adorsys.opba.protocol.bpmnshared.dto.messages.ConsentAcquired;
 import de.adorsys.opba.protocol.xs2a.service.ValidatedExecution;
 import de.adorsys.opba.protocol.xs2a.service.xs2a.RedirectExecutor;
 import de.adorsys.opba.protocol.xs2a.service.xs2a.context.Xs2aContext;
