+++ opba-protocols/xs2a-protocol/src/main/java/de/adorsys/opba/protocol/xs2a/service/xs2a/consent/Xs2aPersistConsentAndContext.java
@@ -2,7 +2,7 @@ package de.adorsys.opba.protocol.xs2a.service.xs2a.consent;
 
 import com.google.common.collect.ImmutableMap;
 import de.adorsys.opba.protocol.api.services.scoped.consent.ProtocolFacingConsent;
-import de.adorsys.opba.protocol.xs2a.config.flowable.Xs2aObjectMapper;
+import de.adorsys.opba.protocol.bpmnshared.config.flowable.Xs2aObjectMapper;
 import de.adorsys.opba.protocol.xs2a.service.ValidatedExecution;
 import de.adorsys.opba.protocol.xs2a.service.xs2a.context.Xs2aContext;
 import lombok.RequiredArgsConstructor;
